Institution Details - toggle additional fields
