+++ src/filing/institutions/constants.js
@@ -1,25 +1,3 @@
-export const institutionTypes = {
-  '-1': 'None chosen',
-  '1': 'National Bank',
-  '2': 'State Member Bank',
-  '3': 'State Non-Member Bank',
-  '4': 'State Chartered Thrift',
-  '5': 'Federal Chartered Thrift',
-  '6': 'Credit Unions',
-  '7': 'Federal Branch or Agency of FBO',
-  '8': 'Branch or Agency of FB',
-  '9': 'MBS of National Bank',
-  '10': 'MBS of State Member Bank',
-  '11': 'MBS of State Non-Member Bank',
-  '12': 'MBS of BHC',
-  '13': 'MBS of CU',
-  '14': 'Independent MBS',
-  '15': 'MBS of S&L HC',
-  '16': 'MBS of State-Chartered Thrift',
-  '17': 'MBS of Federal-Chartered Thrift',
-  '18': 'Affiliate'
-}
-
 export const agencyCodes = {
   1: 'Office of the Comptroller of the Currency (OCC)',
   2: 'Federal Reserve System (FRS)',
